requirements.txt
emoji==1.6.1
fasttext==0.9.2
langchain==0.2.3
matplotlib==3.6.2
news_please==1.5.35
newspaper3k==0.2.8
openai==0.27.4
pandas==1.5.1
paramiko==3.4.0
rank_bm25==0.2.2
tiktoken==0.6.0
torch==2.3.0
tqdm==4.64.1
user_agent==0.1.10
vllm==0.4.2
requests
numpy
langchain-community
langchain-core
